+++ src/index.html
@@ -4,69 +4,63 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="style.css" />
+    <link rel="shortcut icon" href="./img/Totoro1.ico" type="image/x-icon">
     <link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Outfit:wght@100&family=Roboto:wght@100&display=swap" rel="stylesheet">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Outfit:wght@100&family=Roboto:wght@100&display=swap" rel="stylesheet">
     <title>Ghibli Dreams (Data Lovers)</title>
   </head>
-<body>
-  <!-- header -->
-  <header>
-    <img src="./img/banner.png" class="banner" alt="Banner de totoro en fondo turqueza - Ghibli Dreams ">
-    </header>
-     
-      <section class="header">
-        <a class="logo" href="./">
-          <img src="./img/mainimg.png" alt="mainimg">
-          <h2 class="name">Ghibli Dreams</h2>
-        </a>
-         <!-- navegador -->
-      
-          <nav class="menu">
-          <form class="searchForm" action="#" >
-            <input type="text" id="search" name="search" placeholder="Search">
-            <button type="submit">Search</button>
-          </form>
-          </nav>
+  <body >
+    <!-- header -->
+    <section class="header" id="home">
+        <header id="header"><img src="./img/banner.png" class="banner" alt="Banner de totoro en fondo turqueza - Ghibli Dreams ">  </header>  
+    </section>
+      <!-- navegador -->
+      <section class="button">
+        <nav>
+   <!--<a href="#home"><img src="./img/web.png" alt="#home" class="home"></a>-->
+        </nav>
+        <nav>
+        <form class="searchForm" action="#" >
+          <input type="text" id="searchFilm"  name="search" placeholder="Search and press the button" autocomplete="off">
+          <button type="submit">Search a film</button>
+        </form>
+        <section class="btn" id="btn">
+            <label for="" ></label>
+            <select class="btnproducer" id="selectProducer">
+              <option disabled selected>Producer</option>
+              <option value="all" class="allProducer">All</option>
+              <option value="Isao Takahata" class="filterProducer">Isao Takahata</option>
+              <option value="Hayao Miyazaki" class="filterProducer">Hayao Miyazaki</option>
+              <option value="Toru Hara" class="filterProducer">Toru Hara</option>
+              <option value="Toshio Suzuki" class="filterProducer">Toshio Suzuki</option>
+              <option value="Yoshiaki Nishimura" class="filterProducer">Yoshiaki Nishimura</option>
+            </select>
+        </section>
 
-          <nav class="btn" id="btn">
-          
-              <label for="" class="producer"></label>
-              <select class="btnproducer" id="selectProducer">
-                <option disabled selected>Producer</option>
-                <option value="all" class="allProducer">All</option>
-                <option value="Isao Takahata" class="filterProducer">Isao Takahata</option>
-                <option value="Hayao Miyazaki" class="filterProducer">Hayao Miyazaki</option>
-                <option value="Toru Hara" class="filterProducer">Toru Hara</option>
-                <option value="Toshio Suzuki" class="filterProducer">Toshio Suzuki</option>
-                <option value="Yoshiaki Nishimura" class="filterProducer">Yoshiaki Nishimura</option>
-              </select>
+      <section class="btn">
+          <label for="" ></label>
+          <select class="btnOrder" id="selectOrder">
+            <option disabled selected>Order</option>
+            <option value="alphabetical" class="">A - Z</option>
+            <option value="alphabetical" class="alphabetical">Z - A</option>
+          </select>
+        </section>
+  </nav>
 
-          </nav>
-
-          <nav>
-            <label for="" class="order"></label>
-            <select class="btnOrder" id="selectOrder">
-              <option disabled selected>Order</option>
-              <option value="alphabetical" class="">A - Z</option>
-              <option value="alphabetical" class="alphabetical">Z - A</option>
-            </select> 
-            </nav>
-          </section>
-    </nav>
-      <main>
-
-<section class="title">
+  <main class="container">
+    <section class="title">
     <h1> WELCOME TO GHIBLI DREAMS</h1>
-    <p>Studio Ghibli is a renowned Japanese animation studio known for its enchanting and visually stunning films. Here are some of their most beloved works!
-</section>
-
-<section class="root" id="root">
-</section>
-
+    <p>Studio Ghibli is a renowned Japanese animation studio known for its enchanting and visually stunning films. Here are some of their most beloved works.</p>
+    <h2 class="subtitle">Films</h2>
+    </section>
+    
+    <section class="root" id="root">
+    </section>
+  </main>
 <footer>
-  <img class="logoBottom" src="img/Logofoot.png" alt="logo">
-  <p class="copyrightText">Created by Yanet y Yohana</p>
+  <img class="logobottom" src="img/Logofoot.png" alt="logo">
+  <p class="copyright">Created by Yanet & Yohana</p>
   <ul class="socialfooter">
     <li><img src="img/instagram.png" alt="instagram"></li>
     <li><img src="img/twitter.png" alt="twitter"></li>
@@ -74,6 +68,4 @@
 </footer>
 <script src="main.js" type="module"></script>
 </body>
-  
-  </html>
-  
+</html>
