+++ src/style.css
@@ -1,130 +1,75 @@
-body {
+ /*  GENERAL*/
+ body {
   margin: 0;
   padding: 0;
   background-color: #eee;
   font-family: 'Outfit', sans-serif;
   font-family: 'Roboto', sans-serif;
 }
+
 /*header*/
   header {
     text-align: center;
-    width: 100%; 
-  }
 
-  /*navegador*/
+  }
+ /*navegador*/
   .banner
   {  
     width: 100%; 
   }
   nav {
-    
-    color: #fff;
+    background-color: #eaeaea;
+    color: #190808;
     padding: 10px;
-  }
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
 
-  nav  {
-    list-style: none;
-  
-    padding: 0;
-    margin: 0;
-  }
-  
-  nav {
-    display: inline-block;
-    margin-right: 10px;
-  }
-  
-  nav a {
-    display: block;
-    
-    text-align: center;
-    padding: 10px;
-    text-decoration: none;
-  }
-
-  nav form {
-    display: inline-block;
-   
-  }
-  
-  nav input[type="text"] {
-    padding: 6px;
-    border: none;
-    border-radius: 4px;
-    margin-top: 8px;
-  }
-  
-  nav button[type="submit"] {
-    padding: 6px 10px;
-    background-color: #0c0c0c;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  nav label {
-    display: inline-block;
-    margin-right: 10px;
-  }
-  
-  nav select {
-    padding: 6px;
-    border: none;
-    border-radius: 4px;
-    margin-top: 8px;
-    background-color: #f2f2f2;
-  }
-  
-  nav .btn {
-    display: inline-block;
-   
-  }
-    
-  
-  nav a:hover {
-    background-color: #555;
-  }
-  
-
-
-
-
-
-
- 
-  
- /* Estilos para el formulario de búsqueda 
-.searchForm  {
+nav .searchForm {
   display: flex;
   align-items: center;
 }
 
-.searchForm input[type="text"] {
-  padding: 5px;
+nav .searchForm input[type="text"] {
+  padding: 14px 28px;
   border: 1px solid #ccc;
   border-radius: 4px;
   margin-right: 5px;
 }
 
-.searchForm button {
-  padding: 5px 10px;
-  background-color: #333;
+nav .searchForm button {
+  padding: 14px 28px;
+  background-color:#2395D3;;
   color: #fff;
   border: none;
   border-radius: 4px;
   cursor: pointer;
 }
 
-.searchForm button:hover {
-  background-color: #555;
+.btnOrder {
+  padding: 14px 28px;
+  background-color: #fff;
+  color: #333;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
 }
-  
 
+nav .btn {
+  margin-left: 10px;
+}
 
+nav .btn select {
+  padding: 14px 28px;
+  background-color: #fff;
+  color: #333;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
 
-
-/* Titulo */
-  
+/* Titulo y subtítulo */
 h1 {
   font-size: 40px;
   text-align: center;
@@ -132,28 +77,33 @@ h1 {
 }
 
 p {
-  font-size: 30px;
+  font-size: 25px;
   text-align: center;
 }
 
+.subtitle{ 
+  font-size: 30px;
+  color: #eee;
+  text-align: center;
+  margin-bottom: 3rem;
+  box-shadow: 0px 0px 10px 2px rgb(116, 111, 111);
+  background-color:#2395D3;
+}
+
 #root {
-  font-size: 18px;
-  font-weight: bold;
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: center;
-  align-items: flex-start;
-  width: 75%;
-  height: 100%;
-  margin: auto;
-  padding: 10px;
-  margin-bottom: 100px;
-  margin-top: 25px;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    margin-left: 40px;
+    flex-direction: column;
+    margin-bottom: 50px;  
+    font-size: 18px;
+    display: flex;
 }
 
 #root img {
-  width: 17em;
-  height: 25em;
+  width: 25em;
+  height: 35em;
   margin: 15px; 
   border-radius: 4%;
   display: flex;
@@ -165,15 +115,26 @@ p {
 
 #root h2 {
     color: rgb(189, 186, 186);
-    font-size: 1em;
+    font-size: 1.5em;
     width: 90%;
     margin: auto;
     font-weight: bold;
     } 
 
+#root p {
+  color: #052e2f;
+  width: 50vh;
+  display: block;
+  justify-content: flex-start;
+  margin-left: 15px;
+  font-size: 25px;
+  text-align: center;
+}
+
 #root img:hover  {
     position: relative;
-    box-shadow: 0px 0px 30px 10px rgba(44, 44, 44, 0.658);
+    box-shadow: 0px 0px 30px 10px  #78c5ee;;
+    transition: .5s;
     bottom: 15px;
 }
   
@@ -189,14 +150,13 @@ footer {
   padding-top: 3px;
   position: sticky;
   z-index: 15;
-
 }
 
-.logoBottom {
+.logobottom {
   margin-left: 15px;
 }
 
-.copyrightText {
+.copyrighttext {
   text-align: center;
   text-shadow: slategrey;
   margin-top: 40px;
@@ -208,4 +168,3 @@ footer {
   list-style: none;
   filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
 }
-
